+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/BlockNoteEditor.tsx
@@ -1,4 +1,7 @@
-import { BlockNoteEditor as BlockNoteEditorCore } from '@blocknote/core';
+import {
+  BlockNoteEditor as BlockNoteEditorCore,
+  locales,
+} from '@blocknote/core';
 import '@blocknote/core/fonts/inter.css';
 import { BlockNoteView } from '@blocknote/mantine';
 import '@blocknote/mantine/style.css';
@@ -17,6 +20,8 @@ import { useDocStore } from '../stores';
 import { randomColor } from '../utils';
 
 import { BlockNoteToolbar } from './BlockNoteToolbar';
+
+import { useTranslation } from 'react-i18next';
 
 const cssEditor = `
   &, & > .bn-container, & .ProseMirror {
@@ -94,6 +99,19 @@ export const BlockNoteContent = ({
     [createDocAttachment, doc.id],
   );
 
+  const { t, i18n } = useTranslation();
+  const lang = i18n.language;
+
+  const resetStore = () => {
+    setStore(storeId, { editor: undefined });
+  };
+
+  // Invalidate the stored editor when the language changes
+  useEffect(() => {
+    resetStore();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [lang]);
+
   const editor = useMemo(() => {
     if (storedEditor) {
       return storedEditor;
@@ -108,9 +126,10 @@ export const BlockNoteContent = ({
           color: randomColor(),
         },
       },
+      dictionary: locales[lang as keyof typeof locales],
       uploadFile,
     });
-  }, [provider, storedEditor, uploadFile, userData?.email]);
+  }, [provider, storedEditor, uploadFile, userData?.email, lang]);
 
   useEffect(() => {
     setStore(storeId, { editor });